==> scripts/02_metabolite_arafield_transformation.Rmd <==
---
title: "Transformation of metabolite data Arabidopsis field trials"
author: "Heike Sprenger"
date: "July 17, 2016"
output:
  html_document:
    highlight: tango
    number_section: yes
    theme: cerulean
    toc: yes
    toc_float: true
    toc_depth: 4
---

# Set working directory  
```{r set working directory}
# getwd()
# setwd("D:/work/repos/ara_field_trials/")
# setwd("~/work/repos/ara_field_trials/")
# setwd("X:/mpimp/repos/ara_field_trials")
setwd("~/sync/Seafile/uschi-work/repos/ara_field_trials")
```

# Load workspace, packages and scripts
```{r load workspace, include=TRUE, message=FALSE}
# load packages
library(knitr)
library(pander)
library(ggplot2)
library(DAAG)
library(leaps)
library(faraway)
library(MASS)
library(car)

# set options for knitr
opts_chunk$set(fig.width=6, fig.height=5, tidy=TRUE, tidy.opts=list(width.cutoff=70))
opts_knit$set(root.dir = '../')

# load workspace
#load("metabolite_arafield.RData")
```


# Load data
Use changed metabolite data, from ``metabolite_arafield4_overview4.Rmd``
```{r load data}
metabolite_arafield2 <- read.table("output/metabolite_arafield2.txt", header=TRUE, sep="\t")
metabolite_arafield3 <- read.table("output/metabolite_arafield3.txt", header=TRUE, sep="\t")
metabolite_arafield4 <- read.table("output/metabolite_arafield4.txt", header=TRUE, sep="\t")

lt50 <- read.table("data/LT50_64acc.txt", header=TRUE, sep="\t")
dim(lt50) # 64 accessions
pander(head(lt50))

# yCapacity2full <- metabolite_arafield2$capacity
# yCapacity3full <- metabolite_arafield3$capacity
# yCapacity4full <- metabolite_arafield4$capacity
lt50$Capacity <- lt50$LT50ACC - lt50$LT50NAC
```


# Prolin als Beispiel fuer Transformation
## Histogramm und q-q-plot vor Transformation
```{r prolin as example}
pdf("figures/Metabolite_AraField4_Pro_norm_dist.pdf", 9, 5)
par(mfrow=c(1,2))
hist(metabolite_arafield4$proline_positive, freq=FALSE, main="Histogramm", 
     xlab="Konzentration in mg/g FG", ylim=c(0,8),ylab="relative Haeufigkeitsdichte", cex.main=1.3)
lines(density(metabolite_arafield4$proline_positive), col="red")
text(0.3,7, paste("Shapiro-Test: \n P-Wert = \n ",round(shapiro.test(metabolite_arafield4$proline_positive)$p.value, digits=14)), cex=1.1)
qqnorm(metabolite_arafield4$proline_positive,cex.main=1.3, ylab= "tatsaechliche Quantile", xlab="theoretische Quantile")
qqline(metabolite_arafield4$proline_positive)
par(mfrow=c(1,1))
dev.off()

# find optimal transformation parameter lambda
lambda.pro4 <- boxcox(metabolite_arafield4$proline_positive ~ 1)
lambda.max.pro4 <- lambda.pro4$x[which.max(lambda.pro4$y)]
# -0.0606

# --> log-Transformation
bc.pro4 <- bcPower(metabolite_arafield4$proline_positive, lambda = 0)   

# lambda of zero corresponds to log()
# default: natural logarithm
# log(metabolite_arafield4$proline_positive) 

pdf("figures/Metabolite_AraField4_Pro_boxcox.pdf", 5,5)
boxcox(metabolite_arafield4$proline_positive ~ 1)     
title(main="Box-Cox-Plot", cex.main=1.3)
dev.off()

pdf("figures/Metabolite_AraField4_ProT_norm_dist.pdf", 9,5)
par(mfrow=c(1,2))
hist(bc.pro4, freq=FALSE, main="Histogramm", 
     xlab=expression("log"[e]*"(Konzentration in mg/g FG)"), 
     ylim=c(0,0.8),ylab="relative Haeufigkeitsdichte", cex.main=1.3)
lines(density(bc.pro4), col="red")
text(-3.4,0.7, paste("Shapiro-Test: \n P-Wert = \n ",
                     round(shapiro.test(bc.pro4)$p.value, 
                           digits=2)), cex=1.1)
qqnorm(bc.pro4, cex.main=1.3, 
       ylab= "tatsaechliche Quantile", 
       xlab="theoretische Quantile")
qqline(bc.pro4)
par(mfrow=c(1,1))
dev.off()
```


# Histograms to vizualize distribution of metabolite values 
```{r histograms to vizualize distribution of metabolite values}

func_hist <- function(x_values, main_text, xlab_text = "concentration in µmol/g FW", 
                      ylab_text = "relative frequency", text_y = 0.05, ylim_values = NULL){
  hist(x_values, ylim = ylim_values,
       freq = FALSE, main = main_text,
       xlab = xlab_text, ylab = ylab_text,
       cex.main = 1.5, cex.axis = 1.3, cex.lab = 1.3)
  
  lines(density(x_values, na.rm = T), col = "red")
  
  text(summary(x_values)[5], text_y, 
       paste("P-value = \n ", format(shapiro.test(x_values)$p.value, scientific = TRUE, digits = 3)), 
       cex = 1.3)
}
```


## Histograms LT50
```{r histograms LT50}
pdf("figures/LT50_Histogramme.pdf")
par(mar=c(5.1, 5.1, 4.1, 2.1))

func_hist(lt50$LT50NAC, main_text = "freezing tolerance non-acclimated plants",
          xlab_text = expression("LT"[50]*"NAC in °C"))

func_hist(lt50$LT50ACC, main_text = "freezing tolerance acclimated plants",
          xlab_text = expression("LT"[50]*"ACC in °C"))

func_hist(lt50$Capacity, main_text = "acclimation capacity",
          xlab_text = expression("acclimation capacity of LT"[50]*" in °C"))
dev.off()
```


## Histograms Field 2
```{r histograms field 2}
pdf("figures/Metabolite_AraField2_Histogramme.pdf")
par(mar=c(5.1, 5.1, 4.1, 2.1))
func_hist(metabolite_arafield2$glucose_positive, main_text = "glucose", ylim_values = c(0, 0.08))
func_hist(metabolite_arafield2$fructose_positive, main_text = "fructose")
func_hist(metabolite_arafield2$sucrose_positive, main_text = "sucrose")
func_hist(metabolite_arafield2$raffinose_positive, main_text = "raffinose")
func_hist(metabolite_arafield2$amino_acids_positive, main_text = "amino acids")
func_hist(metabolite_arafield2$malate_positive, main_text = "malate")
func_hist(metabolite_arafield2$fumarate_positive, main_text = "fumarate")
func_hist(metabolite_arafield2$proline_positive, main_text = "proline")
dev.off()
```


## Histograms Field 3
```{r histograms field 3}
pdf("figures/Metabolite_AraField3_Histogramme.pdf")
par(mar=c(5.1, 5.1, 4.1, 2.1))
func_hist(metabolite_arafield3$glucose_positive, main_text = "glucose", ylim_values = c(0, 0.08))
func_hist(metabolite_arafield3$fructose_positive, main_text = "fructose")
func_hist(metabolite_arafield3$sucrose_positive, main_text = "sucrose")
func_hist(metabolite_arafield3$raffinose_positive, main_text = "raffinose")
func_hist(metabolite_arafield3$amino_acids_positive, main_text = "amino acids")
func_hist(metabolite_arafield3$malate_positive, main_text = "malate")
func_hist(metabolite_arafield3$fumarate_positive, main_text = "fumarate", text_y = 0.03)
func_hist(metabolite_arafield3$proline_positive, main_text = "proline")
dev.off()
```


## Histograms Field 4
```{r histograms field 4}
pdf("figures/Metabolite_AraField4_Histogramme.pdf")
par(mar=c(5.1, 5.1, 4.1, 2.1))
func_hist(metabolite_arafield4$glucose_positive, main_text = "glucose", ylim_values = c(0, 0.08))
func_hist(metabolite_arafield4$fructose_positive, main_text = "fructose")
func_hist(metabolite_arafield4$sucrose_positive, main_text = "sucrose")
func_hist(metabolite_arafield4$raffinose_positive, main_text = "raffinose")
func_hist(metabolite_arafield4$amino_acids_positive, main_text = "amino acids")
func_hist(metabolite_arafield4$malate_positive, main_text = "malate")
func_hist(metabolite_arafield4$fumarate_positive, main_text = "fumarate")
func_hist(metabolite_arafield4$proline_positive, main_text = "proline")
#removed: func_hist(metabolite_arafield4$anthocyanin_positive, main_text = "anthocyanins")
dev.off()
```


# Shapiro-test
```{r shapiro-test}
# define function for shapiro test
shaptest <- function(n) shapiro.test(n)$p.value

nonNeg_colnames <- c("glucose_positive", "fructose_positive", "sucrose_positive",
                    "raffinose_positive", "amino_acids_positive", "malate_positive", 
                    "fumarate_positive", "proline_positive")

# use nonNeg_colnames
apply(metabolite_arafield4[ , nonNeg_colnames], 2, shaptest)
#     glucose_positive    fructose_positive     sucrose_positive   raffinose_positive 
#         4.737082e-03         6.074753e-05         4.944796e-09         1.953081e-03 
# amino_acids_positive      malate_positive    fumarate_positive     proline_positive 
#         8.937363e-06         5.311129e-01         1.318525e-08         5.412432e-12 

# Shapiro-Test, Nullhypothese: Annahme der Normalverteilung (Ablehnung von H0, wenn p.value<0.05)

shapiro.test(lt50$LT50NAC)$p.value
# 0.1576
shapiro.test(lt50$LT50ACC)$p.value
# 0.0273
shapiro.test(lt50$Capacity)$p.value
# 0.2067
```


# Box-Cox Transformations for Linear Models
**Computes and optionally plots profile log-likelihoods for the parameter of the Box-Cox power transformation.**

## boxcox plots field LT50
```{r boxcox plots LT50}
pdf("figures/LT50_boxcox.pdf")
boxcox(abs(lt50$LT50NAC) ~ 1)
title(main="Box-Cox-Plot LT50NAC", cex.main=1.3)

boxcox(abs(lt50$LT50ACC) ~ 1)
title(main="Box-Cox-Plot LT50ACC", cex.main=1.3)

boxcox(abs(lt50$Capacity) ~ 1)
title(main="Box-Cox-Plot Capacity", cex.main=1.3)
dev.off()
```


## boxcox plots field 2
```{r boxcox plots field 2}
pdf("figures/Metabolite_AraField2_boxcox.pdf")
boxcox(metabolite_arafield4$glucose_positive ~ 1)
title(main="Box-Cox-Plot glucose", cex.main=1.3)

boxcox(metabolite_arafield2$fructose_positive ~ 1)
title(main="Box-Cox-Plot fructose", cex.main=1.3)

boxcox(metabolite_arafield2$sucrose_positive ~ 1)
title(main="Box-Cox-Plot sucrose", cex.main=1.3)

boxcox(metabolite_arafield2$raffinose_positive ~ 1)
title(main="Box-Cox-Plot raffinose", cex.main=1.3)

boxcox(metabolite_arafield2$amino_acids_positive ~ 1)
title(main="Box-Cox-Plot amino acids", cex.main=1.3)

boxcox(metabolite_arafield2$malate_positive ~ 1) 
title(main="Box-Cox-Plot malate", cex.main=1.3)

boxcox(metabolite_arafield2$fumarate_positive ~ 1)
title(main="Box-Cox-Plot fumarate", cex.main=1.3)

boxcox(metabolite_arafield2$proline_positive ~ 1)  
title(main="Box-Cox-Plot proline", cex.main=1.3)
dev.off()
```

## boxcox plots field 3
```{r boxcox plots field 3}
pdf("figures/Metabolite_Arafield3_boxcox.pdf")
boxcox(metabolite_arafield3$glucose_positive ~ 1)
title(main="Box-Cox-Plot glucose", cex.main=1.3)

boxcox(metabolite_arafield3$fructose_positive ~ 1)
title(main="Box-Cox-Plot fructose", cex.main=1.3)

boxcox(metabolite_arafield3$sucrose_positive ~ 1)
title(main="Box-Cox-Plot sucrose", cex.main=1.3)

boxcox(metabolite_arafield3$raffinose_positive ~ 1)
title(main="Box-Cox-Plot raffinose", cex.main=1.3)

boxcox(metabolite_arafield3$amino_acids_positive ~ 1)
title(main="Box-Cox-Plot amino acids", cex.main=1.3)

boxcox(metabolite_arafield3$malate_positive ~ 1)        
title(main="Box-Cox-Plot malate", cex.main=1.3)

boxcox(metabolite_arafield3$fumarate_positive ~ 1)
title(main="Box-Cox-Plot fumarate", cex.main=1.3)

boxcox(metabolite_arafield3$proline_positive ~ 1) 
title(main="Box-Cox-Plot proline", cex.main=1.3)
dev.off()
```

## boxcox plots field 4
```{r boxcox plots field 4}
pdf("figures/Metabolite_AraField4_boxcox.pdf")
boxcox(metabolite_arafield4$glucose_positive ~ 1)
title(main="Box-Cox-Plot glucose", cex.main=1.3)

boxcox(metabolite_arafield4$fructose_positive ~ 1)
title(main="Box-Cox-Plot fructose", cex.main=1.3)

boxcox(metabolite_arafield4$sucrose_positive ~ 1)
title(main="Box-Cox-Plot sucrose", cex.main=1.3)

boxcox(metabolite_arafield4$raffinose_positive ~ 1)
title(main="Box-Cox-Plot raffinose", cex.main=1.3)

boxcox(metabolite_arafield4$amino_acids_positive ~ 1)
title(main="Box-Cox-Plot amino acids", cex.main=1.3)

boxcox(metabolite_arafield4$malate_positive ~ 1)         # lambda = 1 --> keine Transformation noetig
title(main="Box-Cox-Plot malate", cex.main=1.3)

boxcox(metabolite_arafield4$fumarate_positive ~ 1)
title(main="Box-Cox-Plot fumarate", cex.main=1.3)

boxcox(metabolite_arafield4$proline_positive ~ 1)      # lambda = 0 --> log Transformation
title(main="Box-Cox-Plot proline", cex.main=1.3)

# removed 
# boxcox(metabolite_arafield4$anthocyanin_positive ~ 1)
# title(main="Box-Cox-Plot anthocyanins", cex.main=1.3)
dev.off()
```


# Transformation of metabolite values
```{r transformation of metabolite values}
# idx_arafield4 <- with(metabolite_arafield4, which(sucrose_positive<33))
# metabolite_arafield4A <- metabolite_arafield4[idx_arafield4,]

# define function for Box-Cox transformation
func_bc_transform <- function(x_values){
  
  print(paste("p-value before transformation: ", format(shapiro.test(x_values)$p.value, scientific = T)))
  
  lambda_value <- boxcox(x_values ~ 1)
  lambda_max <- lambda_value$x[which.max(lambda_value$y)]
  print(paste("lambda for maximal log-likelihood: ", lambda_max))
  
  x_transformed <- bcPower(x_values, lambda_max)
  print(paste("p-value after transformation: ", format(shapiro.test(x_transformed)$p.value, scientific = T)))
  
  return(x_transformed)
}
```


## Transformation: Field 2
```{r transformation field 2}
metabolite_arafield2$glucoseT <- func_bc_transform(metabolite_arafield2$glucose_positive)
metabolite_arafield2$fructoseT <- func_bc_transform(metabolite_arafield2$fructose_positive)
metabolite_arafield2$sucroseT <- func_bc_transform(metabolite_arafield2$sucrose_positive)
metabolite_arafield2$raffinoseT <- func_bc_transform(metabolite_arafield2$raffinose_positive)
metabolite_arafield2$amino_acidsT <- func_bc_transform(metabolite_arafield2$amino_acids_positive)
metabolite_arafield2$malateT <- func_bc_transform(metabolite_arafield2$malate_positive)
metabolite_arafield2$fumarateT <- func_bc_transform(metabolite_arafield2$fumarate_positive)
metabolite_arafield2$prolineT <- func_bc_transform(metabolite_arafield2$proline_positive)
```


## Transformation: Field 3
```{r transformation field 3}
metabolite_arafield3$glucoseT <- func_bc_transform(metabolite_arafield3$glucose_positive)
metabolite_arafield3$fructoseT <- func_bc_transform(metabolite_arafield3$fructose_positive)
metabolite_arafield3$sucroseT <- func_bc_transform(metabolite_arafield3$sucrose_positive)
metabolite_arafield3$raffinoseT <- func_bc_transform(metabolite_arafield3$raffinose_positive)
metabolite_arafield3$amino_acidsT <- func_bc_transform(metabolite_arafield3$amino_acids_positive)
metabolite_arafield3$malateT <- func_bc_transform(metabolite_arafield3$malate_positive)
metabolite_arafield3$fumarateT <- func_bc_transform(metabolite_arafield3$fumarate_positive)
metabolite_arafield3$prolineT <- func_bc_transform(metabolite_arafield3$proline_positive)
```


## Transformation: Field 4
```{r transformation field 4}
metabolite_arafield4$glucoseT <- func_bc_transform(metabolite_arafield4$glucose_positive)
metabolite_arafield4$fructoseT <- func_bc_transform(metabolite_arafield4$fructose_positive)
metabolite_arafield4$sucroseT <- func_bc_transform(metabolite_arafield4$sucrose_positive)
metabolite_arafield4$raffinoseT <- func_bc_transform(metabolite_arafield4$raffinose_positive)
metabolite_arafield4$amino_acidsT <- func_bc_transform(metabolite_arafield4$amino_acids_positive)
metabolite_arafield4$malateT <- func_bc_transform(metabolite_arafield4$malate_positive) # 1.07 --> keine Transformation noetig
metabolite_arafield4$fumarateT <- func_bc_transform(metabolite_arafield4$fumarate_positive)
metabolite_arafield4$prolineT <- func_bc_transform(metabolite_arafield4$proline_positive)
```


# Plots for transformation of metabolite values
```{r plots for transformation of metabolite values}
# define function for plot of histogram and box-cox-plot of (un)transformed metabolite values
func_plot_transform <- function(x_values, main_text, xlab_text = "concentration in µmol/g FW", 
                                ylab_text = "relative frequency", text_y = 0.05, ylim_values = NULL,
                                x_transformed){
  hist(x_values, ylim = ylim_values,
       freq = FALSE, main = main_text,
       xlab = xlab_text, ylab = ylab_text,
       cex.main = 1.3, cex.axis = 1.3, cex.lab = 1.3)
  
  lines(density(x_values, na.rm = T), col = "red")
  
  text(summary(x_values)[5], text_y, 
       paste("P-value = \n ", format(shapiro.test(x_values)$p.value, scientific = TRUE, digits = 3)), 
       cex = 1.3)
  
  boxCox(x_values ~ 1)
  title(main = "Box-Cox-Plot", cex.main = 1.3)
  
  hist(x_transformed, freq = FALSE, main = "after transformation", xlab = xlab_text, 
       ylab = ylab_text, cex.main=1.3)
  
  lines(density(x_transformed, na.rm = T), col = "red")
  
  text(summary(x_transformed)[5], text_y,
       paste("P-value = \n ", format(shapiro.test(x_transformed)$p.value, scientific = TRUE, digits = 3)),
       cex = 1.3)
  
  qqnorm(x_transformed, cex.main = 1.3)
  qqline(x_transformed)
}
```


## Plots for transformation of metabolite values: Field 2
```{r plots for transformation of metabolite values field 2}
pdf("figures/Metabolite_AraField2_NV.pdf")
par(mfrow=c(2,2))
par(mar=c(5.1, 5.1, 4.1, 2.1))  
func_plot_transform(metabolite_arafield2$glucose_positive, main_text = "glucose", 
                    x_transformed = metabolite_arafield2$glucoseT, ylim_values = c(0, 0.08))
func_plot_transform(metabolite_arafield2$fructose_positive, main_text = "fructose", 
                    x_transformed = metabolite_arafield2$fructoseT)
func_plot_transform(metabolite_arafield2$sucrose_positive, main_text = "sucrose", 
                    x_transformed = metabolite_arafield2$sucroseT)
func_plot_transform(metabolite_arafield2$raffinose_positive, main_text = "raffinose", 
                    x_transformed = metabolite_arafield2$raffinoseT)
func_plot_transform(metabolite_arafield2$amino_acids_positive, main_text = "amino acids", 
                    x_transformed = metabolite_arafield2$amino_acidsT)
func_plot_transform(metabolite_arafield2$malate_positive, main_text = "malate", 
                    x_transformed = metabolite_arafield2$malateT)
func_plot_transform(metabolite_arafield2$fumarate_positive, main_text = "fumarate", 
                    x_transformed = metabolite_arafield2$fumarateT)
func_plot_transform(metabolite_arafield2$proline_positive, main_text = "proline", 
                    x_transformed = metabolite_arafield2$prolineT)
dev.off()
```


## Plots for transformation of metabolite values: Field 3
```{r plots for transformation of metabolite values field 3}
pdf("figures/Metabolite_AraField3_NV.pdf")
par(mfrow=c(2,2))
par(mar=c(5.1, 5.1, 4.1, 2.1))  
func_plot_transform(metabolite_arafield3$glucose_positive, main_text = "glucose", 
                    x_transformed = metabolite_arafield3$glucoseT, ylim_values = c(0, 0.08))
func_plot_transform(metabolite_arafield3$fructose_positive, main_text = "fructose", 
                    x_transformed = metabolite_arafield3$fructoseT)
func_plot_transform(metabolite_arafield3$sucrose_positive, main_text = "sucrose", 
                    x_transformed = metabolite_arafield3$sucroseT)
func_plot_transform(metabolite_arafield3$raffinose_positive, main_text = "raffinose", 
                    x_transformed = metabolite_arafield3$raffinoseT)
func_plot_transform(metabolite_arafield3$amino_acids_positive, main_text = "amino acids", 
                    x_transformed = metabolite_arafield3$amino_acidsT)
func_plot_transform(metabolite_arafield3$malate_positive, main_text = "malate", 
                    x_transformed = metabolite_arafield3$malateT)
func_plot_transform(metabolite_arafield3$fumarate_positive, main_text = "fumarate", text_y = 0.03,
                    x_transformed = metabolite_arafield3$fumarateT)
func_plot_transform(metabolite_arafield3$proline_positive, main_text = "proline", text_y = 0.2, 
                    x_transformed = metabolite_arafield3$prolineT)
dev.off()
```


## Plots for transformation of metabolite values: Field 4
```{r plots for transformation of metabolite values field 4}
pdf("figures/Metabolite_AraField4_NV.pdf")
par(mfrow=c(2,2))
par(mar=c(5.1, 5.1, 4.1, 2.1))  
func_plot_transform(metabolite_arafield4$glucose_positive, main_text = "glucose", 
                    x_transformed = metabolite_arafield4$glucoseT, ylim_values = c(0, 0.08))
func_plot_transform(metabolite_arafield4$fructose_positive, main_text = "fructose", 
                    x_transformed = metabolite_arafield4$fructoseT)
func_plot_transform(metabolite_arafield4$sucrose_positive, main_text = "sucrose", 
                    x_transformed = metabolite_arafield4$sucroseT)
func_plot_transform(metabolite_arafield4$raffinose_positive, main_text = "raffinose", 
                    x_transformed = metabolite_arafield4$raffinoseT)
func_plot_transform(metabolite_arafield4$amino_acids_positive, main_text = "amino acids", 
                    x_transformed = metabolite_arafield4$amino_acidsT)
func_plot_transform(metabolite_arafield4$malate_positive, main_text = "malate", 
                    x_transformed = metabolite_arafield4$malateT)
func_plot_transform(metabolite_arafield4$fumarate_positive, main_text = "fumarate", 
                    x_transformed = metabolite_arafield4$fumarateT)
func_plot_transform(metabolite_arafield4$proline_positive, main_text = "proline", 
                    x_transformed = metabolite_arafield4$prolineT)
dev.off()
```


# Export transformed dataset
```{r export transformed dataset}
write.table(metabolite_arafield2, file="output/metabolite_arafield2_modified.txt", sep="\t")
write.table(metabolite_arafield3, file="output/metabolite_arafield3_modified.txt", sep="\t")
write.table(metabolite_arafield4, file="output/metabolite_arafield4_modified.txt", sep="\t")
```


# Save workspace
```{r save workspace}
save.image("metabolite_arafield.RData")
sessionInfo()
```
